+++ v6/pay.h
@@ -40,7 +40,7 @@ typedef struct {
  PROTOTYPES
  ******************************************************************************/
 
-int pay_read      (void* block, int size, bp_blk_pay_t* pay, bool update_indices, uint16_t* flags);
-int pay_write     (void* block, int size, bp_blk_pay_t* pay, bool update_indices, uint16_t* flags);
+int pay_read      (void* block, int size, bp_blk_pay_t* pay, bool update_indices, uint32_t* flags);
+int pay_write     (void* block, int size, bp_blk_pay_t* pay, bool update_indices, uint32_t* flags);
 
 #endif  /* _pay_h_ */
